LayoutTests/imported/w3c/web-platform-tests/preload/preconnect-onerror-event-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ')'

Harness Error (FAIL), message = SyntaxError: Unexpected token ')'
